refactor: WidgetStore
